Change the log level of the destructor call
The information provided from this message is not useful to anyone in
a production environment. It is incredibly spammy, as it would trigger
on every instantiated object, even if unused. It'll trash your logs
with this message for every destroyed instance.

Considering the force_bulk is called in the destructor anyway, I'd
judge DEBUG to be the correct level for this, as even INFO doesn't fit
it.